Refactor get_tick_bars function to accept a float threshold

+++ pdr_backend/lake/alt_bar.py
@@ -143,7 +143,7 @@ def get_volume_bars(rawohlcv_df: pd.DataFrame, threshold: float):
     return list_bars, newest_ut_value
 
 
-def get_tick_bars(rawohlcv_df: pd.DataFrame, threshold: int):
+def get_tick_bars(rawohlcv_df: pd.DataFrame, threshold: float):
     """
     Creates the tick bars: date_time, open, high, low, close, cum_vol, cum_dollar, and cum_ticks.
